imunita za univerzální nápovědy/c-kódy

+++ src/Repo/HintRepo.php
@@ -1,6 +1,7 @@
 <?php
 namespace Sova\Repo;
 
+use \PDO;
 use Sova\DBException;
 
 class HintRepo extends RepoBase {
@@ -17,21 +18,27 @@ class HintRepo extends RepoBase {
 		}
 	}
 
-	public function getUnusedHintCount($teamId) {
-		return $this->db->equery("SELECT COUNT(*) FROM hint WHERE team_id = ? AND cipher_id IS NULL", $teamId);
+	public function getUnusedHintCount(int $teamId) {
+		return $this->db->equery("SELECT COUNT(*) FROM hint WHERE team_id = ? AND type IS NULL", $teamId);
 	}
 
-	public function getUnusedHint($teamId) {
-		return $this->db->squery("SELECT * FROM hint WHERE team_id = ? AND cipher_id IS NULL LIMIT 1", $teamId);
+	public function getUnusedHint(int $teamId, int $count = 1) {
+		return $this->db->aquery("SELECT * FROM hint WHERE team_id = ? AND type IS NULL LIMIT ?", $teamId, $count);
 	}
 
 	public function alreadyApplied(array $hint) {
 		return $this->db->equery("SELECT COUNT(*) FROM hint WHERE team_id = :team_id AND cipher_id = :cipher_id AND time <= NOW()", $hint) != 0;
 	}
 
+	public function imunityApplied(int $teamId) {
+		return $this->db->equery("SELECT COUNT(*) FROM hint WHERE team_id = ? AND type = 3", $teamId) != 0;
+	}
+
 	public function apply(array $hint) {
-		$this->db->execute("DELETE FROM hint WHERE team_id = :team_id AND cipher_id = :cipher_id AND type = :type AND time > NOW()", $hint);
-		$this->db->execute("UPDATE hint SET cipher_id = :cipher_id, time = NOW(), type = :type WHERE team_id = :team_id AND unihint_id = :unihint_id", $hint);
+		if ($hint["type"] != 3) {
+			$this->db->execute("DELETE FROM hint WHERE team_id = :team_id AND cipher_id = :cipher_id AND type = :type AND time > NOW()", $hint);
+		}
+		$this->db->execute("UPDATE hint SET cipher_id = :cipher_id, time = NOW(), type = :type WHERE hint_id = :hint_id", $hint);
 	}
 
 	public function amend(array &$hint) {
